Demo 5: Statistical Analysis of Techniques
This is a rework of Demo 4 to allow the experiment to be repeated an arbitrary number of times, in order to enable statistical analysis of this approach vs linear brute force.

The algorithm runtime might be slower due to the cycle-detection requiring two scalar mults vs the linear search only requiring one, but note that the overall number of iterations is lower for the PoC than the linear strategy.